給付実績テーブル　Entity最新化対応

+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/KyufuJissekiCareManagementHi.java
@@ -258,7 +258,7 @@ extends ModelBase<KyufuJissekiCareManagementHiIdentifier,
      *
      * @return 単位数
      */
-    public Decimal get単位数() {
+    public int get単位数() {
         return entity.getTanisu();
     }
 
@@ -267,7 +267,7 @@ extends ModelBase<KyufuJissekiCareManagementHiIdentifier,
      *
      * @return 回数
      */
-    public Decimal get回数() {
+    public int get回数() {
         return entity.getKaisu();
     }
 
@@ -330,7 +330,7 @@ extends ModelBase<KyufuJissekiCareManagementHiIdentifier,
      *
      * @return 後_単位数
      */
-    public Decimal get後_単位数() {
+    public int get後_単位数() {
         return entity.getAtoTanisu();
     }
 
@@ -339,7 +339,7 @@ extends ModelBase<KyufuJissekiCareManagementHiIdentifier,
      *
      * @return 後_回数
      */
-    public Decimal get後_回数() {
+    public int get後_回数() {
         return entity.getAtoKaisu();
     }
 
@@ -384,7 +384,7 @@ extends ModelBase<KyufuJissekiCareManagementHiIdentifier,
      *
      * @return 再審査回数
      */
-    public Decimal get再審査回数() {
+    public int get再審査回数() {
         return entity.getSaishinsaKaisu();
     }
 
@@ -393,7 +393,7 @@ extends ModelBase<KyufuJissekiCareManagementHiIdentifier,
      *
      * @return 過誤回数
      */
-    public Decimal get過誤回数() {
+    public int get過誤回数() {
         return entity.getKagoKaisu();
     }
