+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/KyufuJissekiShakaiFukushiHojinKeigengakuBuilder.java
@@ -255,7 +255,7 @@ public class KyufuJissekiShakaiFukushiHojinKeigengakuBuilder {
      * @param 再審査回数 再審査回数
      * @return {@link KyufuJissekiShakaiFukushiHojinKeigengakuBuilder}
      */
-    public KyufuJissekiShakaiFukushiHojinKeigengakuBuilder set再審査回数(Decimal 再審査回数) {
+    public KyufuJissekiShakaiFukushiHojinKeigengakuBuilder set再審査回数(int 再審査回数) {
         requireNonNull(再審査回数, UrSystemErrorMessages.値がnull.getReplacedMessage("再審査回数"));
         entity.setSaishinsaKaisu(再審査回数);
         return this;
@@ -267,7 +267,7 @@ public class KyufuJissekiShakaiFukushiHojinKeigengakuBuilder {
      * @param 過誤回数 過誤回数
      * @return {@link KyufuJissekiShakaiFukushiHojinKeigengakuBuilder}
      */
-    public KyufuJissekiShakaiFukushiHojinKeigengakuBuilder set過誤回数(Decimal 過誤回数) {
+    public KyufuJissekiShakaiFukushiHojinKeigengakuBuilder set過誤回数(int 過誤回数) {
         requireNonNull(過誤回数, UrSystemErrorMessages.値がnull.getReplacedMessage("過誤回数"));
         entity.setKagoKaisu(過誤回数);
         return this;
